Add basic iOS support
Note: iOS has two targets.
- devices (arm64)
- simulators (arm64, x86_64)

It's more complex to make a single library.
So for development if needed, there is the flag `WEBGPU_IOS_SIMULATOR`.
With this flag it will be possible to compile for simulator,
but this will fail for real devices so it's needs to be re-configured
for each task.